fetch community post within one year
